Fix dependency between install and config.
Previously, the main config file depended on the install class, but the
other resources in the config dir didn't.  This sometimes led to errors
when attempting to create the directories etc, because the owning user
doesn't exist until the package is installed.  Making the whole config
class require the install class resolves this.